Providing a ready-to-use example in README.md (#39)
* Providing a ready-to-use example in README.md

I think an introductory example in the README.md of a project should be both self-explaining and self-containing (i.e. not requiring additional code) if possible. That wasn't the case.
I tried to accomplish this by adapting the existing code.
I also adapted the label of the link to the full to do app.

* fix construction

* improve console output

(at least for Chrome)

* fix typo